Move all mapping in MappingProfile
