+++ backend/CircleForms/Models/Mapping/MappingProfile.cs
@@ -1,15 +1,21 @@
 ﻿using AutoMapper;
+using CircleForms.Models.Posts;
+using CircleForms.Models.Posts.Questions;
+using CircleForms.Models.Posts.Questions.Submissions;
 
-namespace CircleForms.Models.Posts.Questions.Submissions;
+namespace CircleForms.Models.Mapping;
 
-public class SubmissionContractToSubmissionMapper : Profile
+public class MappingProfile : Profile
 {
-    public SubmissionContractToSubmissionMapper()
+    public MappingProfile()
     {
         CreateMap<(QuestionType type, SubmissionContract contract), Submission>()
             .ForMember(x => x.QuestionId, x => x.MapFrom(v => v.contract.QuestionId))
             .ForMember(x => x.Value,
                 o => o.MapFrom(src => CalculateAnswer(src)));
+
+        CreateMap<Post, PostRedis>()
+            .ForMember(x => x.Id, x => x.MapFrom(v => v.Id.ToString()));
     }
 
     private static string CalculateAnswer((QuestionType type, SubmissionContract contract) value)
