--- backend/CircleForms/Models/Posts/PostToPostRedisMapper.cs
@@ -1,12 +1,0 @@
-﻿using AutoMapper;
-
-namespace CircleForms.Models.Posts;
-
-public class PostToPostRedisMapper : Profile
-{
-    public PostToPostRedisMapper()
-    {
-        CreateMap<Post, PostRedis>()
-            .ForMember(x => x.Id, x => x.MapFrom(v => v.Id.ToString()));
-    }
-}
